# Performance optimization of STL loading