--- a/src/compas/_iotools.py
+++ b/src/compas/_iotools.py
@@ -1,11 +1,42 @@
 """For the time being, these functions are only for internal use."""
-from contextlib import contextmanager
-import io
+from __future__ import print_function
 
-try:
+import io
+import platform
+from contextlib import contextmanager
+
+if "ironpython" == platform.python_implementation().lower():
+    from System import Array
+    from System import Byte
+    from System.IO import MemoryStream
+    from System.Net import SecurityProtocolType
+    from System.Net import ServicePointManager
+    from System.Net import WebRequest
+
+    def download_url_as_bytes(url):
+        ServicePointManager.SecurityProtocol = SecurityProtocolType.Tls12
+
+        request = WebRequest.Create(url)
+
+        response = request.GetResponse()
+        responseStream = response.GetResponseStream()
+
+        memoryStream = MemoryStream()
+        buffer = Array.CreateInstance(Byte, 8192)
+        bytesRead = responseStream.Read(buffer, 0, buffer.Length)
+
+        while bytesRead > 0:
+            memoryStream.Write(buffer, 0, bytesRead)
+            bytesRead = responseStream.Read(buffer, 0, buffer.Length)
+
+        return memoryStream.ToArray()
+
+else:
     from urllib.request import urlopen
-except ImportError:
-    from urllib2 import urlopen
+
+    def download_url_as_bytes(url):
+        response = urlopen(url)
+        return response.read()
 
 
 @contextmanager
@@ -45,8 +76,7 @@ def open_file(file_or_filename, mode="r"):
             # support all read modes (r, r+, rb, etc)
             if not mode.startswith("r"):
                 raise ValueError("URLs can only be opened in read mode.")
-            response = urlopen(file)
-            file = io.BytesIO(response.read())
+            file = io.BytesIO(download_url_as_bytes(file))
         else:
             file = open(file, mode=mode)
             close_source = True
@@ -77,3 +107,63 @@ def iter_file(file, size=65536):
         if not data:
             break
         yield data
+
+
+if __name__ == "__main__":
+    import time
+
+    from compas.datastructures import Mesh
+
+    platform.RANDOM = "python"
+    iters = 5.0
+
+    if "ironpython" == platform.python_implementation().lower():
+        print("Running on urllib2 implementation")
+        for i in range(5):
+            all_times = []
+            for j in range(int(iters)):
+                t1 = time.time()
+                url = r"https://raw.githubusercontent.com/ros-industrial/abb/kinetic-devel/abb_irb6600_support/meshes/irb6640/visual/link_{}.stl".format(
+                    i + 1
+                )
+                mesh = Mesh.from_stl(url)
+                if i > 0:  # discard first iteration
+                    all_times.append(time.time() - t1)
+                    print("  File {}, took {:.3f} secs".format(i, all_times[-1]), end=", ")
+            if all_times:
+                print()
+                print("Average: {:.3f} secs".format(sum(all_times) / iters))
+
+        print("Switch to .NET implementation")
+        platform.RANDOM = ".NET"
+        for i in range(5):
+            all_times = []
+            for j in range(int(iters)):
+                t1 = time.time()
+                url = r"https://raw.githubusercontent.com/ros-industrial/abb/kinetic-devel/abb_irb6600_support/meshes/irb6640/visual/link_{}.stl".format(
+                    i + 1
+                )
+                mesh = Mesh.from_stl(url)
+                if i > 0:  # discard first iteration
+                    all_times.append(time.time() - t1)
+                    print("  File {}, took {:.3f} secs".format(i, all_times[-1]), end=", ")
+            if all_times:
+                print()
+                print("Average: {:.3f} secs".format(sum(all_times) / iters))
+
+    else:
+        print("Running on urllib.request implementation")
+        for i in range(5):
+            all_times = []
+            for j in range(int(iters)):
+                t1 = time.time()
+                url = r"https://raw.githubusercontent.com/ros-industrial/abb/kinetic-devel/abb_irb6600_support/meshes/irb6640/visual/link_{}.stl".format(
+                    i + 1
+                )
+                mesh = Mesh.from_stl(url)
+                if i > 0:  # discard first iteration
+                    all_times.append(time.time() - t1)
+                    print("  File {}, took {:.3f} secs".format(i, all_times[-1]), end=", ", flush=True)
+            if all_times:
+                print()
+                print("Average: {:.3f} secs".format(sum(all_times) / iters))